Initial renaming of files
