+++ ros_package_template/include/ros_package_template/algorithm.hpp
