+++ ros_package_template/include/ros_package_template/ros_package_template.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "ros_package_template/Algorithm.hpp"
+#include "ros_package_template/algorithm.hpp"
 
 // ROS
 #include <ros/ros.h>
