+++ ros_package_template/src/algorithm.cpp
@@ -1,4 +1,4 @@
-#include "ros_package_template/Algorithm.hpp"
+#include "ros_package_template/algorithm.hpp"
 
 namespace ros_package_template {
